Remove unnecessary indirection
